requirements-train.txt
# =============================================================================
# Training dependencies — Rehab AI (Gesture + Brain Scan pipelines)
# Install: pip install -r requirements-train.txt
# =============================================================================

# --- Core ML (tabular / gesture pipeline: train_optimized.py) ---------------
scikit-learn>=1.3.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# --- Brain scan image pipeline (train_brain_scan.py) ------------------------
# PyTorch CPU build (use CUDA URL if you have a GPU):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0

# --- Real-time system (realtime.py) -----------------------------------------
mediapipe>=0.10.0
opencv-python>=4.8.0

# --- ESP32 serial comms (optional) ------------------------------------------
pyserial>=3.5
